Changed GFI event helper functions to GfiHelper.js
